+++ lib/src/widgets/question_tab.dart
@@ -41,7 +41,7 @@ class QuestionTab extends StatelessWidget {
                   question.titleShort != ""
               ? '${question.titleShort} ${question.question}'
               : question.question,
-          style: Theme.of(context).textTheme.headlineMedium,
+          style: Theme.of(context).textTheme.headlineLarge,
         ),
         SizedBox(
           height: 10.0,